Add files via upload
In this module, you have started to get acquainted with the C++ programming language. We propose to take a very important step in mastering this language — to write your first program. You will have the opportunity to work with the real code and solve the first tasks. At the end of this longrid, you will find hints on completing the task and an example with answers, but we encourage you to complete the tasks yourself first, and only then watch the hint.